uftrace: Convert to getopt_long() for option parsing
The argp_parse() has portability issues as some systems miss the
implementation by default.  So convert it to getopt_long() for better
compatibility.  It needs more manual work when you add a new option,
but it'd be more flexible to handle help messages at least.

The notable change is that it now requires subcommand name as the
first argument.  This is somewhat annoying for me but I don't know how
I can fix it with getopt().

So the following example will not work anymore (sadly).

  $ ./uftrace -L . record tests/t-abc

It should be like below:

  $ ./uftrace record -L . tests/t-abc

Closed: #871

Signed-off-by: Namhyung Kim <namhyung@gmail.com>