mcount: Fix crash when shmem accessed after task is dead
When a task is dead, shmem_finish() is called and shmem->buffer is
freed.  After that some functions might be called for some reason.
While get_shmem_buffer() has some code to protect such condition it
was not perfect and caused a NULL pointer dereference due to
shmem->buffer.  I found it during an unit test.

Signed-off-by: Namhyung Kim <namhyung@gmail.com>